dev-portfolio-server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==3.0.10
python-dotenv==1.0.0
